Address code review issues that are simple to fix.

+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/businessrules/StreamingJobRequestBo.java
@@ -96,7 +96,7 @@ public interface StreamingJobRequestBo {
 
 	void handleJobStatusChange(long jobId, JobStatus status, long timestamp);
 
-	void handleNewActivityAlert(long jobId, int frameId, long timestamp);
+	void handleNewActivityAlert(long jobId, long frameId, long timestamp);
 
 	void handleNewSummaryReport(SegmentSummaryReport summaryReport);
 }
